test: add test for bad_allocs during large mv queries
This patch adds a test for reproducing issue #12379, which is
being fixed in #16819.
The test case works by creating a large table with multiple
materialized views, and then performing a partition delete
query for all partitions. This used to result in a bad_alloc
caused by too many view updates being generated concurrently,
the test passes if all queries are successfully executed.
The test requires multiple nodes, because with only one node
the updates are performed synchronously and the concurrency
is limited.