table: add_sstable_and_update_cache: trigger compaction only in compaction group
There is no need to trigger compaction in all compaction
groups when an sstable is added to only one of them.

And with that level of control, if the caller passes
sstables::offstrategy::yes, we know it will
trigger offstrategy compaction later on so there
is no need to trigger compaction at all
for this sstable at this time.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>