invalidate mv prepared statements when base table changes
This bug was discovered as part of another bug investigation,
if the base table changes, for example a row is added or removed,
the prepared statements of the materialized views doesn't change,
this will result in returning incomplete information until the prepared
statements cache is invalidated for some other reason or until the node
is rebooted.
This patch fixes it by accounting for the fact that a materialized view
select query also depends on it's base table.

Fixes #16392

Signed-off-by: Eliran Sinvani <eliransin@scylladb.com>