test/cql-pytest: add "--tablets" option to test/cql-pytest/run
Adding the "--tablets" option to test/cql-pytest/run simply runs Scylla
with the "tablets" experimental option. When this option is turned on,
Scylla will by default use tablets - instead of vnodes - in all
keyspaces created by all tests.

After this patch, some tests will fail when running with "--tablets",
e.g., because they test for features that we haven't yet implemented
with tablets. This will be fixed in the next patch.

In the future, as tablets mature, turning on the tablets experimental
option will become the default, and a "--vnodes" options will run tests
without tablets. Reversing the logic will be trivial.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>