test/cql-pytest: add "fails_with_tablets" marker
When running tests with tablets enabled (test/cql-pytest/run --tablets),
some tests are expected to fail because of known issues - notably features
not yet implemented with tablets (e.g., CDC).

This patch introduces a fixture "fails_with_tablets" which says that
this test is known to fail when run on tablets. This fixture checks
if running against Scylla with the tablets experimental feature (it doesn't
matter who started this Scylla or how), and if it is, the test is marked
"xfail", i.e., it behave like normal xfail (the test still runs, and
may XPASS if the bug was fixed).

Again, this fixture works correctly no matter if pytest is run against
a manually-run Scylla, or if you use the test/cql-pytest/run or test.py
scripts to run it.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>