test/cql-pytest: disable tablets in Scylla-only schema altering tests
Some tests create a NetworkTopologyStrategy table, and then ALTER the
strategy. If tablets are enabled, we don't allow such a change to occur
because it would mean switching an existing table from tablets to
vnodes, which we don't yet support.

Because these are all Scylla-only tests and create their own keyspace,
it's fine to just disable tablets in those keyspaces to keep these tests
working like they did in the past.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>