test/cql-pytest: fix local storage test to work with tablets
We have a test that checks that "local" storage is the default and
does nothing, but the way this is checked is roundabout and basically
wrong - it checks that nothing gets written in "scylla_keyspaces"
for this keyspace. But when tablets are enabled, something DOES appear
there. So we need to disable tablets for this test to make it work
without changing its logic.

It's fine to disable tablets (a Scylla-only extension) because this
test is already scylla_only.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>