test/cql-pytest: fails_with_tablet for one Cassandra ALTER test
One of the tests in the translated Cassandra unit test alter_test.py
fails with tablets (test/cql-pytest/run --tablets) because it tries
to ALTER a keyspace with NetworkTopologyStrategy (which has tablets
by default) to SimpleStrategy (which doesn't have tablets), and this
is currently not allowed.

I could have fixed it in several ways, but in the spirit of not papering
over places where we really differ in behavior from Cassandra, I prefer
to leave this test as XFAIL when running with tablets. So added the
"fails_with_tablets" fixture to do that.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>