streaming: Optionally disable off-strategy for TWCS tables
off-strategy on TWCS tables has 100% space overhead.

The overhead might prevent scale out if disk usage of a single
TWCS table is above the 50%.

Disabling offstrategy limits the space overhead to 1 / N, where
N is the number of configured windows. The disadvantage is
increased write ampl, but the option is there to help critical
scenarios where the space overhead will cause nodes to run out
of space during node op. Optimizations that temporarily increase
the min threshold (during node op or high write rate) help
mitigating the write ampl problem.

The long term plan is to bring incremental compaction to TWCS,
and after that, the usage of this option can be deprecated.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>