wait_for_ip: make sure the host_id <-> ip mapping is persisted
In this commit we issue an explicit update_peer_info to persist
host_id <-> ip mapping during wait_for_ip command on each node.
Otherwise, if the node crashes and restarts, it may end up in
a state where a new node has successfully joined, but the restarted
node doesn't have an IP for it. We'll include an explicit test
for this scenario in the next commit.

The disadvantage of this approach is that if the join process fails (e.g.
if some other node didn't receive the IP of the joining node in time)
our system.peers table will have a non-existent IP address. This
seems like a minor issue, and could be addressed in the future if
the topology_state_load function learns to reconcile the
system.peers table based on system.topology and add/remove
nodes accordingly.