test/cql-pytest: add parameter support to test_keyspace
Tests can now request to be run against both tablets and vnodes, via:

    @pytest.mark.parametrize("test_keyspace", ["tablets", "vnodes"], indirect=True)

This will set request.param for the test_keyspace fixture, which can
create the keyspace according to the requested parameter. This way,
tests can conveniently opt-in to be run against both replication
methods.
When not parameterized like this, the test_keyspace fixture will create
a keyspace as before -- with tablets, if support is enabled.